Fixed Set/Clear Feature requests directed to a non-configured endpoint not returning a stall to the host.
Make TeensyHID bootloader use an 8 byte control endpoint to match the official Teensy, just in case Paul decides to authenticate on that attribute in the future.

+++ LUFA/Drivers/USB/LowLevel/Template/Template_Endpoint_Control_W.c
@@ -21,7 +21,7 @@ uint8_t TEMPLATE_FUNC_NAME (const void* Buffer, uint16_t Length)
 		  
 		if (Endpoint_IsINReady())
 		{
-			uint16_t BytesInEndpoint = Endpoint_BytesInEndpoint();
+			uint8_t BytesInEndpoint = Endpoint_BytesInEndpoint();
 		
 			while (Length && (BytesInEndpoint < USB_ControlEndpointSize))
 			{
